[fails] Introduce random-movement bike lane building
This commit introduces the function which chooses whether to build a bike lane
or remove one (decide_building). It changes edit_network (algorithm.py) such
that the mode (rev) of adding or removing bike lanes is switched according to
decide_building. This comes with the drawback that there are no facilities to
stop the simulation right now, so this commit causes the simulation to either
crash because the logging system cannot deal with switching rev, or loop
indefinitely. Normal behavior might be gained by setting total_budget to 0.