[draft] Add "continue computation" feature
The first while-loop in run_algorithm does not depend on build_mode.
So it should be possible to start the computation for different build_modes
based on the same initial computation with the same minmode.

This commit is a draft because this approach does not work yet, as it does
not initialize nkG, nxG and edge_dict (and possibly others). One solution
might be to run the while-loop, but skip all re-calculations of bike paths
and edge selection. Chosen bike paths might need to be recalculated at the
end of the while loop.